Add UIViewController spinner extension for loading
Create UIActivityIndicatorView extension to be used whenever
a screen needs to display a loading spinner.
