+++ StickySessions/StickySessions/View/Login/LoginViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 import Firebase
 import GoogleSignIn
 
-class LoginViewController: UIViewController, GIDSignInUIDelegate {
+class LoginViewController: UIViewController, GIDSignInUIDelegate, GIDSignInDelegate {
 
     static let SEGUE_ID: String = "listSessionsShowNextSegueId"
     var firebaseAuthStateListener: AuthStateDidChangeListenerHandle?
@@ -19,6 +19,7 @@ class LoginViewController: UIViewController, GIDSignInUIDelegate {
         super.viewDidLoad()
 
         GIDSignIn.sharedInstance().uiDelegate = self
+        GIDSignIn.sharedInstance().delegate = self
 
         self.navigationItem.setHidesBackButton(true, animated: false)
     }
@@ -35,6 +36,7 @@ class LoginViewController: UIViewController, GIDSignInUIDelegate {
             if let user = user {
                 print ("user = " + (user.displayName ?? ""))
                 self.loadNextViewController()
+                self.removeSpinner()
             }
         }
     }
@@ -45,6 +47,7 @@ class LoginViewController: UIViewController, GIDSignInUIDelegate {
     
     @IBAction func googleSignIn(_ sender: Any) {
         GIDSignIn.sharedInstance().signIn()
+        self.showSpinner(onView: self.view)
     }
 
     func loadNextViewController() {
@@ -55,6 +58,31 @@ class LoginViewController: UIViewController, GIDSignInUIDelegate {
         super.viewWillDisappear(animated)
         Auth.auth().removeStateDidChangeListener(firebaseAuthStateListener!)
         navigationController?.setNavigationBarHidden(false, animated: animated)
+    }
+
+    func sign(_ signIn: GIDSignIn!, didSignInFor user: GIDGoogleUser!, withError error: Error?) {
+
+        if let error = error {
+            print(error)
+            self.removeSpinner()
+            return
+        }
+
+        guard let authentication = user.authentication else { return }
+        let credential = GoogleAuthProvider.credential(withIDToken: authentication.idToken, accessToken: authentication.accessToken)
+
+        Auth.auth().signIn(with: credential) { (authResult, error) in
+            if let error = error {
+                print(error)
+                self.removeSpinner()
+                return
+            }
+        }
+    }
+
+    func sign(_ signIn: GIDSignIn!, didDisconnectWith user: GIDGoogleUser!, withError error: Error!) {
+        // Perform any operations when the user disconnects from app here.
+        // ...
     }
 
 }
